test(domain): separate benefit-related tests of EventPlannerTest
